requirements.txt
pygame
luma.lcd
luma.core
Pillow
opencv-python
evdev
